feedback_service/requirements.txt
fastapi[standard]==0.124.0
uvicorn[standard]==0.38.0

sqlmodel==0.0.27
asyncpg==0.31.0
psycopg[binary]==3.3.2
alembic==1.17.2

pydantic-settings==2.12.0
email-validator==2.3.0

python-jose[cryptography]==3.5.0

aiokafka==0.12.0
cramjam==2.11.0

httpx==0.28.1
structlog==24.4.0

# ── Voice / Audio pipeline ────────────────────────────────────────────────
# Object storage (MinIO / S3-compatible)
aiobotocore==2.15.2

# File I/O helpers
aiofiles==24.1.0
python-multipart==0.0.20

# STT: Whisper API is called via httpx (already listed above).
# For self-hosted faster-whisper (WHISPER_MODE=local), install separately
# in a custom Docker image — not included here to keep the default image small.
# faster-whisper==1.1.0   # uncomment + add to Dockerfile if using local mode

# TTS: Google / Azure / ElevenLabs called via httpx (already listed above).
# No additional SDK needed for REST-based TTS providers.
